% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{BartlettTestOfSphericity}
\alias{BartlettTestOfSphericity}
\title{\code{BartlettTestOfSphericity}}
\usage{
BartlettTestOfSphericity(data, weights = NULL, subset = NULL,
  missing = "Exclude cases with missing data")
}
\arguments{
\item{data}{A data frame containing the data to test.}

\item{weights}{A numeric vector containing the weight for each case in data.}

\item{subset}{A logical vector which describes the subset of \code{data} to
be analyzed.}

\item{missing}{A string specifiying what to do when the \code{data} contains
missing values. The valid options are \code{"Error if missing data"},
\code{"Exclude cases with missing data"}, \code{"Use partial data (pairwise
correlations)"}, and \code{"Imputation (replace missing values with
estimates)"}.}
}
\value{
A list containing the Chi-Square value, degrees of freedom
  (\code{df}), and p-value for the test.
}
\description{
Conduct the Bartlett Test of Sphericity for a set of data, which
  tests that the correlation matrix of the data is not the identity matrix.
}
\details{
This function wraps \code{\link[psych]{cortest.bartlett}}. In
  particular, it extends the existing funcitonality to weighted data, and it
  computes the test using a more conservative value of the sample size when
  there is missing data. The value for the sample size that is used is the
  size of the smallest pairwise-complete set of cases among all pairs of
  variables. This is consistent with SPSS.
}

