% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{CANormalization}
\alias{CANormalization}
\title{\code{CANormalization}}
\usage{
CANormalization(ca.object, normalization = "Principal")
}
\arguments{
\item{ca.object}{The object to normalize.}

\item{normalization}{The method used to normalize the coordinates of the
correspondence analysis plot (this changes the plot, but not the outputs of
\code{\link[ca]{ca}} itself. The default method is \code{"Principal"},
which plots the principal coordinates (i.e., the standard coordinates
multipled by the singular values). \code{"Row principal"} and \code{"Column
principal"} plot the standard coordinates of the columns (rows) against the
principal coordinates. \code{"Symmetrical (\u00BD)"} plots the standard
coordinates multiplied by the square root of the singular values.
\code{"None"} plots the standard coordinates.}
}
\description{
Normalizes the coordinates of a \code{\link[ca]{ca}} object.
}

