% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsanalysis.R
\name{ComponentPlot}
\alias{ComponentPlot}
\title{\code{ComponentPlot}}
\usage{
ComponentPlot(x, show.labels = TRUE)
}
\arguments{
\item{x}{An object of class \code{flipFactorAnalysis}.}

\item{show.labels}{Label the points with the row names.}
}
\description{
Create a scatter plot showing the loadings of each variable on
  the first two principal components.
}
