% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{ConvertVariablesForFactorAnalysis}
\alias{ConvertVariablesForFactorAnalysis}
\title{\code{ConvertVariablesForFactorAnalysis}}
\usage{
ConvertVariablesForFactorAnalysis(variables)
}
\arguments{
\item{variables}{A list of variables that you want to include in your analysis.}
}
\description{
Convert variables to the appropriate form for use in PCA or Factor Analysis
}
\details{
Numeric variables are unchanged. Ordered factors are replaced with a numeric
variable containing their levels. Non-ordered factors are converted to a set of indicator
(binary) variables with one variable for each level but the first.
}

