% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{CorrespondenceAnalysis}
\alias{CorrespondenceAnalysis}
\title{\code{CorrespondenceAnalysis}}
\usage{
CorrespondenceAnalysis(x, normalization = "Principal",
  output = c("Scatterplot", "Bubble Chart", "Moonplot", "Text")[1],
  row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"), row.color = "#5B9BD5",
  col.color = "#ED7D31", bubble.size = NULL, bubble.title = "",
  chart.title = "Correspondence analysis", transpose = FALSE,
  logos = NULL, logo.size = 0.5, ...)
}
\arguments{
\item{x}{A table.}

\item{normalization}{The method used to normalize the coordinates of the
correspondence analysis plot (this changes the plot, but not the outputs of
\code{\link[ca]{ca}} itself. The default method is \code{"Principal"},
which plots the principal coordinates (i.e., the standard coordinates
multiplied by the singular values). \code{"Row principal"} and \code{"Column
principal"} plot the standard coordinates of the columns (rows) against the
principal coordinates. Note that the plotting occurs via
\code{\link{print.CorrespondenceAnalysis}}.}

\item{output}{How the map is displayed: \code{"Scatterplot"}, or \code{"Moonplot"}, or \code{"Text"}.}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.
variable is provided, any cases with missing values on this variable are
excluded from the final data file.}

\item{row.color}{Color to display row-attributes in scatterplot.}

\item{col.color}{Color to display column-attributes in scatterplot.}

\item{bubble.size}{A vector of magnitudes for the row coordinate (for bubble charts). This is optional.}

\item{bubble.title}{A label for the legend.}

\item{chart.title}{Title of chart.}

\item{transpose}{Boolean indicating whether the rows and columns of \code{x} should be swapped.}

\item{logos}{Optional list of images to be used to label scatterplot instead of the row names. It should be inputted as a comma-seperated list of URLs.}

\item{logo.size}{Numeric controlling the size of the logos.}

\item{...}{Optional arguments for \code{\link[ca]{ca}}.}
}
\description{
Removes rows or columns from the table.
}
\details{
Where a matrix or array is passed in containing names for the dimensions, these are used to represent the rows
and columns in the legend. If there are no names, then the names are assumed to be the contents of \code{attr(x, "row.column.names")}.
If there are still no names, they are assumed to be \code{Rows} and \code{Columns}, respectively.
}
