% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancematrix.R
\name{DistanceMatrix}
\alias{DistanceMatrix}
\title{\code{DistanceMatrix}}
\usage{
DistanceMatrix(input.data, compare = "Variables", case.labels = NULL,
  variable.names = FALSE, binary = FALSE,
  measure = "Dissimilarities", similarity.measure = "Correlation",
  distance.measure = "Euclidian", minkowski = 2,
  standardization = "None", standardize.by = "Case",
  measure.transformation = "None", show.cell.values = "Automatic",
  show.row.labels = "Yes", show.column.labels = "Yes", subset = NULL,
  weights = NULL)
}
\arguments{
\item{input.data}{A \code{\link{data.frame}} with cases as rows and variables as columns.}

\item{compare}{Whether \code{"Variables"} or \code{"Cases"} are to be compared.}

\item{case.labels}{If \code{"compare"} is set to \code{"Cases"}, a list of labels for each case. If not supplied
the indices are used.}

\item{variable.names}{Whether to display variable names or labels.}

\item{binary}{Whether to convert categorical variables to binary.}

\item{measure}{Whether to measure \code{"Similarities"} or \code{"Dissimilarities"}.}

\item{similarity.measure}{How to calculate similarities. Options are \code{"Correlation"}
or \code{"Cosine"}.}

\item{distance.measure}{The method for mesuring distances. Options are \code{"Euclidean"},
\code{"Squared Euclidean"}, \code{"Maximum"}, \code{"Manhattan"} and \code{"Minkowski"}.}

\item{minkowski}{The power used by \code{"Minkowski"} \code{"distance.measure"}.}

\item{standardization}{The method used to standardize the data. Options are \code{"None"},
\code{"z-scores"}, \code{"Range [-1,1]"}, \code{"Range [0,1]"}, \code{"Mean of 1"} and
\code{"Standard deviation of 1"}.}

\item{standardize.by}{Whether to perform \code{"standardization"} by \code{"Variable"} or
by \code{"Case"}.}

\item{measure.transformation}{How to transform the measure. Options are \code{"None"},
\code{"Absolute values"}, \code{"Reverse sign"}, and \code{"Range [0,1]"}.}

\item{show.cell.values}{Whether to show cell values on heatmap output. Choices are
\code{"Automatic"}, \code{"Yes"} and \code{"No"}.}

\item{show.row.labels}{\code{"Yes"} or \code{"No"}.}

\item{show.column.labels}{\code{"Yes"} or \code{"No"}.}

\item{subset}{An optional vector specifying a subset of cases to be used.}

\item{weights}{An optional vector of sampling weights.}
}
\description{
Compute a distance (dissimilarity) or similarity matrix between variables or cases.
}
