% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{fit.mcaObj}
\alias{fit.mcaObj}
\title{\code{fit.mca}}
\usage{
fit.mcaObj(object, keep.missing = FALSE)
}
\arguments{
\item{object}{A mca object created using \code{MultipleCorrespondenceAnalysis}.}
}
\description{
Computes coordinates from multiple correspondance analysis
}

