% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionreductionscatterplot.R
\name{fitted.2Dreduction}
\alias{fitted.2Dreduction}
\title{\code{fitted.2Dreduction}}
\usage{
\method{fitted}{2Dreduction}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"2Dreduction"}.}

\item{...}{Not used.}
}
\description{
\code{fitted.2Dreduction}
}
