% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{fitted.mcaObj}
\alias{fitted.mcaObj}
\title{\code{fitted.mcaObJ}}
\usage{
\method{fitted}{mcaObj}(object, ...)
}
\arguments{
\item{object}{A mca object created using \code{MultipleCorrespondenceAnalysis}.}
}
\description{
Computes projected coordinates from multiple correspondance analysis
}

