% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensionalscaling.R
\name{fitted.MDS}
\alias{fitted.MDS}
\title{\code{fitted.MDS}}
\usage{
\method{fitted}{MDS}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"MDS"}.}

\item{...}{Not used.}
}
\description{
\code{fitted.MDS}
}
