% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{fitted.tSNE}
\alias{fitted.tSNE}
\title{\code{fitted.tSNE}}
\usage{
\method{fitted}{tSNE}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"tSNE"}.}

\item{...}{Not used.}
}
\description{
\code{fitted.tSNE}
}
