% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffitplot.R
\name{GoodnessOfFitPlot.2Dreduction}
\alias{GoodnessOfFitPlot.2Dreduction}
\title{Goodness-of-fit plot for a 2Dreduction object}
\usage{
\method{GoodnessOfFitPlot}{2Dreduction}(object, max.points = 1000, ...)
}
\description{
Goodness-of-fit plot for a 2Dreduction object
}
