% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffitplot.R
\name{GoodnessOfFitPlot.flipFactorAnalysis}
\alias{GoodnessOfFitPlot.flipFactorAnalysis}
\title{Goodness-of-fit plot for a flipFactorAnalysis object}
\usage{
\method{GoodnessOfFitPlot}{flipFactorAnalysis}(object, max.points = 1000,
  ...)
}
\description{
Goodness-of-fit plot for a flipFactorAnalysis object
}
