% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensionalscaling.R
\name{MultiDimesnsionalScaling}
\alias{MultiDimesnsionalScaling}
\title{\code{MultiDimesnsionalScaling}}
\usage{
MultiDimesnsionalScaling(distance.matrix, metric = TRUE)
}
\arguments{
\item{distance.matrix}{A \code{\link{matrix}} of distances between points.}

\item{metric}{A \code{\link{factor}} used to group cases. Ignored if \code{is.distance} is TRUE.}
}
\description{
Perform Multidimesnsional Scaling to produce a 2-dimensional embedding of
a distance matrix.
}
