% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{MultipleCorrespondenceAnalysis}
\alias{MultipleCorrespondenceAnalysis}
\title{\code{MultipleCorrespondanenceAnalysis}}
\usage{
MultipleCorrespondenceAnalysis(formula, data = NULL,
  output = c("Scatterplot", "Moonplot", "Text"), weights = NULL,
  subset = NULL, missing = "Exclude cases with missing data",
  auxiliary.data = NULL, show.labels = FALSE)
}
\arguments{
\item{data}{A data frame containing factor variables.}

\item{output}{Specify output generated. May be one of \code{"Scatterplot", "Moonplot"} or \code{"Text"}.}

\item{weights}{A numeric vector containing the weight for each case in \code{data}.}

\item{subset}{A logical vector which describes the subset of \code{data} to be analysed.}

\item{missing}{A string specifiying what to do when the data contains missing values. This should be one of \code{"Error if missing data", "Exclude cases with missing data"}, and \code{"Imputation (replace missing values with estimates)"}.}
}
\description{
Calculate multiple correspondence analysis of categorical variables
}

