% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{plot.mcaObj}
\alias{plot.mcaObj}
\title{\code{plot.mcaObj}}
\usage{
\method{plot}{mcaObj}(x, ...)
}
\arguments{
\item{x}{The multiple correspondance analysis object to be analysed}

\item{...}{Not used}
}
\description{
Plots scatterplot of two largest principal coordinates from MCA analysis
}

