% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{plot_mcaObj}
\alias{plot_mcaObj}
\title{\code{plot_mcaObj}}
\usage{
plot_mcaObj(x)
}
\arguments{
\item{object}{The multiple correspondance analysis object to be analysed}
}
\description{
Plots object even if output is set to \code{"Text"}
}

