% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{predict.mca}
\alias{predict.mca}
\title{\code{predict.mca}}
\usage{
\method{predict}{mca}(object, newdata = NULL)
}
\arguments{
\item{object}{A mjca object created using \code{MultipleCorrespondenceAnalysis}.}

\item{newdata}{A data frame containing factor variables. Levels must match the factors used to construct the model}
}
\description{
Predict coordinates of new data with multiple correspondance analysis
}

