% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{predict.mcaObj}
\alias{predict.mcaObj}
\title{\code{predict.mca}}
\usage{
\method{predict}{mcaObj}(object, newdata = NULL)
}
\arguments{
\item{object}{A mca object created using \code{MultipleCorrespondenceAnalysis}.}

\item{newdata}{A data frame containing factor variables. Levels must match the factors used to construct the model}
}
\description{
Predict coordinates of new data with multiple correspondance analysis
}

