% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsanalysis.R
\name{prepareDataForFactorAnalysis}
\alias{prepareDataForFactorAnalysis}
\title{\code{prepareDataForFactorAnalysis}}
\usage{
prepareDataForFactorAnalysis(data, weights, subset, missing)
}
\arguments{
\item{data}{A data frame with numeric columns which contains the data to be
analyzed.}

\item{weights}{A numeric vector containing the weight for each case in data.}

\item{subset}{A logical vector which describes the subset of \code{data} to
be analyzed.}

\item{missing}{A string specifiying what to do when the \code{data} contains
missing values. The valid options are \code{"Error if missing data"},
\code{"Exclude cases with missing data"}, \code{"Use partial data (pairwise
correlations)"}, and \code{"Imputation (replace missing values with
estimates)"}.}
}
\value{
A list containing \code{subset.data}, which is a data frame which has had subset applied
and missing values removed or imputed as specified by the parameter \code{missing}, and \code{prepared.weights}
which is a nuneric vector containing the weight values that correspond to the remaining cases (or NULL when
the input weight is NULL).
}
\description{
Filter data, remove cases with missing values, and impute where requested.
}

