% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoranalysis.R
\name{PrincipalComponentsAnalysis}
\alias{PrincipalComponentsAnalysis}
\title{\code{PrincipalComponentsAnalysis}}
\usage{
PrincipalComponentsAnalysis(data, weights = NULL, subset = NULL,
  missing = "Exclude cases with missing data", use.correlation = TRUE,
  rotation = "none", oblimin.delta = 0, promax.kappa = 4,
  select.n.rule = "Number of factors", eigen.min = 1, n.factors = 1,
  sort.coefficients.by.size = FALSE, suppress.small.coefficients = FALSE,
  min.display.loading.value = 0.1, print.type = "loadings",
  show.labels = TRUE, plot.labels = TRUE)
}
\arguments{
\item{data}{A data frame with numeric columns which contains the data to be
analyzed.}

\item{weights}{A numeric vector containing the weight for each case in data.}

\item{subset}{A logical vector which describes the subset of \code{data} to
be analyzed.}

\item{missing}{A string specifiying what to do when the \code{data} contains
missing values. The valid options are \code{"Error if missing data"},
\code{"Exclude cases with missing data"}, \code{"Use partial data (pairwise
correlations)"}, and \code{"Imputation (replace missing values with
estimates)"}.}

\item{use.correlation}{A logical value specifying whether to use the
correlation matrix (\code{TRUE}), or the covariance matrix (\code{FALSE}).}

\item{rotation}{A string specifying the type of rotation to be used. Valid
options are  \code{"none"}, \code{"varimax"}, \code{"quartimax"},
\code{"equamax"}, \code{"promax"}, and \code{"oblimin"}.}

\item{oblimin.delta}{A parameter supplied for oblimin rotations.}

\item{promax.kappa}{A parameter supplied for promax rotations.}

\item{select.n.rule}{Method for selecting the number of principal components to keep. May be one of \code{"Kaiser rule"}, \code{"Eigenvalues over"}, or \code{"Number of components"}.}

\item{eigen.min}{Cut-off above which eigenvalues are selected. Used if \code{select.n.rule} is \code{"Eigenvalues over"}.}

\item{n.factors}{An integer specifying the number of principal components to keep. Used if \code{select.n.rule} is \code{"Number of components"}.}

\item{sort.coefficients.by.size}{A logical value determining whether loadings
should be sorted when printed.}

\item{suppress.small.coefficients}{A logical value specifying whether components
that are less than \code{min.display.loading.value} in magnitude will be replaced
with blanks when printing.}

\item{min.display.loading.value}{Loadings smaller than this value will not be
displayed when printed.}

\item{print.type}{A string specifying the type of printing that should be
done. Valid options are \code{"loadings"} to display a (rotated) loading table,
\code{"structure"} to display a component structure matrix (which is the loadings
multiplied by the component correlations),
\code{"scree"} to display a Scree Plot, and \code{"scatter"} to display a
plot of the first two dimensions of the final loadings. The latter two
options make use of HTML widgets.}

\item{show.labels}{If \code{TRUE}, labels are shown rather than name in outputs.}

\item{plot.labels}{A logical value which determines whether or not the
scatter plot will show the labels of the input data, or just integers
specifying the column number of each variable.}
}
\description{
Calculate a Principal Component Analysis
}
\details{
This uses \code{\link[psych]{principal}} from package \code{psych} to compute the unrotated
PCA, and uses package \code{GPArotation} to find a rotated solution if required, to match SPSS' PCA. The
rotation includes a Kaiser normalization and a method of Promax which matches what SPSS does.
Includes handling of missing data, weighting, and filtering.
}

