% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsbiplot.R
\name{PrincipalComponentsBiplot}
\alias{PrincipalComponentsBiplot}
\title{\code{PrincipalComponentsBiplot}}
\usage{
PrincipalComponentsBiplot(x, normalization = c("Principal", "Row principal",
  "Column principal", "Symmetrical", "None")[1], output = c("Scatterplot",
  "Moonplot", "Text")[1], row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"))
}
\arguments{
\item{x}{A table containing rows and columns#'}

\item{output}{Specify output generated. May be one of \code{"Scatterplot"} or \code{"Text"}.}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.
variable is provided, any cases with missing values on this variable are
excluded from the final data file.}
}
\description{
Performs PCA on a table and outputs biplot
}
\examples{
x2 <- cbind(Glance=c(0.016, 0.058, 0.061, 0.038, 0.01),
           FairlyThorough=c(0.022, 0.147, 0.093, 0.128, 0.022),
           VeryThorough=c(0.006,0.064,0.125,0.157,0.051))
rownames(x2) <- c("SomePrimary","PrimaryCompleted","SomeSecondary",
                  "SecondaryCompleted","SomeTertiary")
PrincipalComponentsBiplot(x2, output="Moonplot")

}

