% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionreductionscatterplot.R
\name{print.2Dreduction}
\alias{print.2Dreduction}
\title{\code{print.2Dreduction}}
\usage{
\method{print}{`2Dreduction`}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"2Dreduction"}.}

\item{...}{Not used.}
}
\description{
If 2Dreduction object is created from distance matrix, then print as labelled scatterplot.
If created from a \code{\link{data.frame}} of variables, print as a scatterplot of grouped cases.
}
