% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{print.CorrespondenceAnalysis}
\alias{print.CorrespondenceAnalysis}
\title{\code{print.CorrespondenceAnalysis}}
\usage{
\method{print}{CorrespondenceAnalysis}(x, ...)
}
\arguments{
\item{x}{CorrespondenceAnalysis object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Creates a plot displaying the correspondence analysis results.
}

