% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancematrix.R
\name{print.DistanceMatrix}
\alias{print.DistanceMatrix}
\title{\code{print.DistanceMatrix}}
\usage{
\method{print}{DistanceMatrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"DistanceMatrix"}.}

\item{...}{Not used.}
}
\description{
Display distance matrix as a heatmap.
}
