% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{print.mca}
\alias{print.mca}
\title{\code{print.mca}}
\usage{
\method{print}{mca}(object)
}
\arguments{
\item{object}{The multiple correspondance analysis object to be analysed}
}
\description{
\code{print.mca}
}

