% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{print.mcaObj}
\alias{print.mcaObj}
\title{\code{print.mcaObj}}
\usage{
\method{print}{mcaObj}(object)
}
\arguments{
\item{object}{The multiple correspondance analysis object to be analysed}
}
\description{
\code{print.mcaObj}
}

