% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensionalscaling.R
\name{print.MDS}
\alias{print.MDS}
\title{\code{print.MDS}}
\usage{
\method{print}{MDS}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"MDS"}.}

\item{...}{Not used.}
}
\description{
\code{print.MDS}
}
