% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsbiplot.R
\name{print.PCAbiplot}
\alias{print.PCAbiplot}
\title{\code{print.PCAbiplot}}
\usage{
\method{print}{PCAbiplot}(x, ...)
}
\arguments{
\item{x}{An object created using \code{PrincipalComponentsBiplot}.}

\item{...}{Not used}

\item{color1}{A string specifying the color of the row-attributes in the scatterplot}

\item{color2}{A string specifying the color of the column-attributes in the scatterplot}
}
\description{
Plots biplot of PCA analysis, showing both the component loadings and scores simultaneously.
}

