% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R, R/goodnessoffitplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ExtractChartData}
\alias{GoodnessOfFitPlot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flipFormat}{\code{\link[flipFormat]{ExtractChartData}}}

  \item{flipStatistics}{\code{\link[flipStatistics]{GoodnessOfFitPlot}}}
}}

