% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsanalysis.R
\name{ScreePlot}
\alias{ScreePlot}
\title{\code{ScreePlot}}
\usage{
ScreePlot(x, weights = NULL, subset = NULL,
  missing = "Exclude cases with missing data", use.correlation = TRUE)
}
\arguments{
\item{x}{Either a data frame, a numeric vector of eigenvalues, or the
eigenvalues from an analysis of class \code{flipFactorAnalysis} from
\code{\link{PrincipalComponentsAnalysis}}, or \code{fa} or \code{principal} from package
psych. When x is a data frame, additional arguments can be supplied as to
how to compute the covariance or correlation matrix.}

\item{weights}{A numeric vector containing the weight for each case in data.}

\item{subset}{A logical vector which describes the subset of \code{data} to
be analyzed.}

\item{missing}{A string specifiying what to do when the \code{data} contains
missing values. The valid options are \code{"Error if missing data"},
\code{"Exclude cases with missing data"}, \code{"Use partial data (pairwise
correlations)"}, and \code{"Imputation (replace missing values with
estimates)"}.}

\item{use.correlation}{A logical value specifying whether to use the
correlation matrix (\code{TRUE}), or the covariance matrix (\code{FALSE}).}
}
\value{
An HTML widget object from plotly containing the Scree Plot.
}
\description{
Plot the eigenvalues from an existing principal component or
  factor analysis or plot the eigenvalues from the correlation or covariance
  matrix of a data frame.
}

