% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tSNE}
\alias{tSNE}
\title{\code{tSNE}}
\usage{
tSNE(data, subset = NULL, data.groups = NULL, is.distance = FALSE,
  binary = TRUE, perplexity = 10, seed = 1066)
}
\arguments{
\item{data}{A \code{\link{data.frame}} or \code{\link{matrix}} containing the data to be analyzed.}

\item{subset}{A logical vector which describes the subset of \code{data} to be analyzed.
Ignored if \code{is.distance} is TRUE.}

\item{data.groups}{A \code{\link{factor}} used to group cases. Ignored if \code{is.distance} is TRUE.}

\item{is.distance}{Whether \code{data} are a distance \code{\link{matrix}} between points or a
\code{\link{data.frame}} of cases by row and features by column.}

\item{binary}{If \code{TRUE}, unordered factors are converted to dummy variables. Otherwise,
they are treated as sequential integers. Ignored if \code{is.distance} is FALSE.}

\item{perplexity}{The perplexity coefficient which defines the extent of the locality
of the dimension reduction.}

\item{seed}{Random seed.}
}
\description{
Produce a t-Distributed Stochastic Neighbour Embedding.
}
\details{
If input is not a distance matrix, any case with missing data or a missing label is ignored.
}
