% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{WeightedTable}
\alias{WeightedTable}
\title{\code{WeightedTable}}
\usage{
WeightedTable(..., weights = NULL, exclude = if (useNA == "no") c(NA, NaN),
  useNA = c("no", "ifany", "always"), dnn = list.names(...),
  deparse.level = 1)
}
\arguments{
\item{...}{one or more objects which can be interpretated as factors, or a list or dataframe whose components can be so interpreted}

\item{exclude}{levels to remove for all factors in \code{...}}

\item{useNA}{whether to include \code{NA} values in the table}

\item{dnn}{the names to given to the dimensions in the result}

\item{deparse.level}{controls how the default \code{dnn} is constructed.}
}
\description{
Generalisation of the \code{table} function in base to handle weights
}

