% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{CAQuality}
\alias{CAQuality}
\title{\code{CAQuality}}
\usage{
CAQuality(ca.object)
}
\arguments{
\item{ca.object}{The object to compute quality for.}
}
\description{
Quality measures of a correspondence analysis.
}
