% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondenceanalysis.R
\name{CorrespondenceAnalysis}
\alias{CorrespondenceAnalysis}
\title{\code{CorrespondenceAnalysis}}
\usage{
CorrespondenceAnalysis(x, normalization = "Principal",
  output = c("Scatterplot", "Bubble Chart", "Moonplot", "Text",
  "Input Table")[1], focus = NULL, supplementary = NULL,
  row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"),
  mirror.horizontal = FALSE, mirror.vertical = FALSE,
  color.palette = "Default colors", row.color = "#5B9BD5",
  col.color = "#ED7D31", bubble.size = NULL, bubble.title = "",
  chart.title = "Correspondence analysis", transpose = FALSE,
  logos = NULL, logo.size = 0.5, trend.lines = FALSE,
  multiple.tables = NA, square = FALSE, max.row.labels.plot = 200,
  max.col.labels.plot = 200, max.labels.plot = NA, dim1.plot = 1,
  dim2.plot = 2, ...)
}
\arguments{
\item{x}{A \code{\link{matrix}} or a list of matrices.}

\item{normalization}{The method used to normalize the coordinates of the
correspondence analysis plot (this changes the plot, but not the outputs of
\code{\link[ca]{ca}} itself. The default method is \code{"Principal"},
which plots the principal coordinates (i.e., the standard coordinates
multiplied by the singular values). \code{"Row principal"} and \code{"Column
principal"} plot the standard coordinates of the columns (rows) against the
principal coordinates. \code{"Row principal (scaled)"} is the same as \code{"Row principal"}
except that both column coordinates are equally scaled so that column points appear
on a similar scale to row points. \code{"Column principal (scaled)"} is analagous.
Note that the plotting occurs via \code{\link{print.CorrespondenceAnalysis}}.}

\item{output}{How the map is displayed: \code{"Scatterplot"}, or \code{"Moonplot"}, or \code{"Text"}.}

\item{focus}{The label of a row or column category. The output is rotated
so that the variance of this category lies along the first dimension.}

\item{supplementary}{A vector of rows or columns to be treated as supplementary, i.e. not included
in the calculation of the coordinate space but to be plotted.}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.}

\item{mirror.horizontal}{Boolean indicating whether to reverse the sign of values plotted along the horizontal axis.}

\item{mirror.vertical}{Boolean indicating whether to reverse the sign of values plotted along the vertical axis.}

\item{color.palette}{Palette used to color scatterplot when multiple tables are used.}

\item{row.color}{Color to display row-attributes in scatterplot with one table.}

\item{col.color}{Color to display column-attributes in scatterplot with one table.}

\item{bubble.size}{A vector of magnitudes for the row coordinate (for bubble charts). This is optional.}

\item{bubble.title}{A label for the legend.}

\item{chart.title}{Title of chart.}

\item{transpose}{Boolean indicating whether the rows and columns of \code{x} should be swapped.}

\item{logos}{Optional list of images to be used to label scatterplot instead of the row names. It should be inputted as a comma-seperated list of URLs.}

\item{logo.size}{Numeric controlling the size of the logos.}

\item{trend.lines}{Boolean indicating whether to draw trend lines when multiple tables are supplied.}

\item{multiple.tables}{Optional boolean indicating whether or not multiple tables have been supplied.
If no value is given, it will be guessed from the structure of \code{x}.}

\item{square}{Boolean indicating whether the input table is square. If true the row and column names of the table must be the same.}

\item{max.row.labels.plot}{A number specifying the maximum of row labels shown in bubble or scatterplots. The remaining rows will be shown with labels hidden.}

\item{max.col.labels.plot}{A number specifying the maximum number column labels shown.}

\item{max.labels.plot}{Deprecated. Use max.row.labels.plot instead.}

\item{dim1.plot}{Dimension to show in X-axis of bubble or scatterplot.}

\item{dim2.plot}{Dimension to show in Y-axis of bubble or scatterplot.}

\item{...}{Optional arguments for \code{\link[ca]{ca}}.}
}
\description{
Removes rows or columns from the table.
}
