% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionreductionscatterplot.R
\name{DimensionReductionScatterplot}
\alias{DimensionReductionScatterplot}
\title{\code{DimensionReductionScatterplot}}
\usage{
DimensionReductionScatterplot(
  algorithm,
  data = NULL,
  data.groups = NULL,
  table = NULL,
  raw.table = FALSE,
  subset = NULL,
  perplexity = 10,
  binary = TRUE,
  normalization = FALSE,
  seed = 1066,
  use.combined.scatter = FALSE,
  ...
)
}
\arguments{
\item{algorithm}{Valid options are \code{"t-SNE"},
\code{"MDS - Metric"}, \code{"MDS - Non-metric"}, or
\code{"PCA"}. Note that \code{'PCA'} does not accept a
distance matrix as input.}

\item{data}{A \code{\link{data.frame}} with cases by row and
variables by column.}

\item{data.groups}{A \code{\link{vector}} to be used as a grouping
variable for the embedded \code{data}.}

\item{table}{A symmetrical distance \code{\link{matrix}} or an
object of class \code{dist} or \code{DistanceMatrix}.}

\item{raw.table}{If \code{TRUE},
\code{\link[flipTransformations]{ParseUserEnteredTable}} is
called on \code{table} to create a numeric matrix from text.}

\item{subset}{A logical vector which describes the subset of
\code{data} to be analyzed.  Not used for \code{table} input.}

\item{perplexity}{The perplexity coefficient which defines the
extent of the locality of the dimension reduction. Used only
when \code{algorithm} is \code{t-SNE}.}

\item{binary}{If \code{TRUE}, unordered factors are converted to
dummy variables. Otherwise, their value attributes/levels are
used to coerce each factor to a single numeric variable. See
the Details of \code{\link{AsNumeric}}.}

\item{normalization}{If \code{data} is supplied, whether to
standardize the data so each variable has a mean of 0 and
standard deviation of 1.}

\item{seed}{Random seed. Used only when \code{algorithm} is
\code{"t-SNE"}.}

\item{use.combined.scatter}{Draw scatterplots using rhtmlCombinedScatter.}

\item{...}{Other parameters passed to
\link{PrincipalComponentsAnalysis}.}
}
\description{
Produces as 2-dimensional embedding. Takes either a
    \code{\link{data.frame}} of variables and optionally another
    \code{\link{vector}} to be used as a grouping variable, or
    takes a distance \code{\link{matrix}}.
}
\details{
For \code{data} input, all algorithms apart from
    \code{PCA} remove duplicated data and any case with \code{NA}
    is ignored by all algorithms.
}
