% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionreductionscatterplot.R
\name{DimensionReductionScatterplot}
\alias{DimensionReductionScatterplot}
\title{\code{DimensionReductionScatterplot}}
\usage{
DimensionReductionScatterplot(algorithm, data = NULL, data.groups = NULL,
  table = NULL, raw.table = FALSE, subset = NULL, perplexity = TRUE,
  binary = TRUE)
}
\arguments{
\item{algorithm}{TODO}

\item{data}{TODO}

\item{data.groups}{TODO}

\item{table}{TODO}

\item{raw.table}{TODO So we know whether to parse user input table. Cols contain labels.}

\item{subset}{A logical vector which describes the subset of \code{data} to
be analyzed. Not used for \code{table} input.}

\item{perplexity}{TODO}

\item{binary}{TODO}
}
\description{
Reduces input to 2 dimensions. Takes either data.frame of variables and
another variable of data.groups, to produce a Scatterplot.  Or takes a distance matrix
to produce a labelled scatterplot.
}
