% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffitplot.R
\name{GoodnessOfFitPlot.flipFactorAnalysis}
\alias{GoodnessOfFitPlot.flipFactorAnalysis}
\title{Goodness-of-fit plot for a flipFactorAnalysis object}
\usage{
\method{GoodnessOfFitPlot}{flipFactorAnalysis}(object, max.points = 1000,
  ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{max.points}{The maximum numner of points to plot.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Goodness-of-fit plot for a flipFactorAnalysis object
}
