% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessoffitplot.R
\name{GoodnessOfFitPlot}
\alias{GoodnessOfFitPlot}
\alias{GoodnessOfFitPlot.default}
\alias{GoodnessOfFitPlot.lm}
\alias{GoodnessOfFitPlot.flipFactorAnalysis}
\alias{GoodnessOfFitPlot.2Dreduction}
\title{\code{GoodnessOfFitPlot}}
\usage{
GoodnessOfFitPlot(object, ...)

\method{GoodnessOfFitPlot}{default}(object, digits = max(3L,
  getOption("digits") - 3L), max.points = 1000, ...)

\method{GoodnessOfFitPlot}{lm}(object, digits = max(3L, getOption("digits") -
  3L), max.points = 1000, ...)

\method{GoodnessOfFitPlot}{flipFactorAnalysis}(object, max.points = 1000, ...)

\method{GoodnessOfFitPlot}{2Dreduction}(object, max.points = 1000, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}

\item{digits}{Minimal number of significant digits, see \code{\link{print.default}}.}

\item{max.points}{The maximum numner of points to plot.}
}
\description{
A generic function used to produce plots illustrating the goodness-of-fit of
the model object.  The function invokes particular \code{\link{methods}}
which depend on the \code{\link{class}} of the first argument.

Reports the goodness-of-fit of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default goodness-of-fit plot

\item \code{lm}: Goodness-of-fit plot for a linear model

\item \code{flipFactorAnalysis}: Goodness-of-fit plot for a flipFactorAnalysis object

\item \code{2Dreduction}: Goodness-of-fit plot for a 2Dreduction object
}}

