% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponentsbiplot.R
\name{PrincipalComponentsBiplot}
\alias{PrincipalComponentsBiplot}
\title{\code{PrincipalComponentsBiplot}}
\usage{
PrincipalComponentsBiplot(x, normalization = c("Principal", "Row principal",
  "Column principal", "Symmetrical", "None")[1], output = c("Scatterplot",
  "Moonplot", "Text")[1], row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"), row.color = "#5B9BD5",
  col.color = "#ED7D31")
}
\arguments{
\item{x}{A table containing rows and columns}

\item{normalization}{Method used to standarize coordinates of biplot. The default method is \code{"Principal"},
which plots the principal coordinates (i.e., the standard coordinates
multiplied by the singular values). \code{"Row principal"} and \code{"Column
principal"} plot the standard coordinates of the columns (rows) against the
principal coordinates.}

\item{output}{Specify output generated. May be one of \code{"Scatterplot", "Moonplot"} or \code{"Text"}.}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.}

\item{row.color}{Color to display row-attributes in scatterplot.}

\item{col.color}{Color to display column-attributes in scatterplot.}
}
\description{
Performs PCA on a table and outputs biplot
}
\details{
Where a matrix or array is passed in containing names for the dimensions, these are used to represent the rows
and columns in the legend. If there are no names, then the names are assumed to be the contents of \code{attr(x, "row.column.names")}.
If there are still no names, they are assumed to be \code{Rows} and \code{Columns}, respectively.
}
\examples{
x2 <- cbind(Glance=c(0.016, 0.058, 0.061, 0.038, 0.01),
           FairlyThorough=c(0.022, 0.147, 0.093, 0.128, 0.022),
           VeryThorough=c(0.006,0.064,0.125,0.157,0.051))
rownames(x2) <- c("SomePrimary","PrimaryCompleted","SomeSecondary",
                  "SecondaryCompleted","SomeTertiary")
PrincipalComponentsBiplot(x2, output="Moonplot")

}
