% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplecorrespondenceanalysis.R
\name{print.mcaObj}
\alias{print.mcaObj}
\title{\code{print.mcaObj}}
\usage{
\method{print}{mcaObj}(x, digits = 3, ...)
}
\arguments{
\item{x}{The multiple correspondance analysis object to be analysed.}

\item{digits}{Integer indicating number of decimal places to be used.}
}
\description{
\code{print.mcaObj}
}

