% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tSNE}
\alias{tSNE}
\title{\code{tSNE}}
\usage{
tSNE(data, subset = NULL, data.labels = NULL, algorithm = "Rtsne",
  binary = TRUE)
}
\arguments{
\item{data}{A \code{\link{data.frame}} which contains the data to be analyzed.}

\item{subset}{A logical vector which describes the subset of \code{data} to be analyzed.}

\item{data.labels}{A \code{\link{vector}} of labels for each case.}

\item{algorithm}{Which package to use. Valid options are \code{"Rtsne"} and \code{"tsne"}.}

\item{binary}{If \code{TRUE}, unordered factors are converted to dummy variables. Otherwise,
they are treated as sequential integers.}
}
\description{
Produce a t-Distributed Stochastic Neighbour Embedding
}
\details{
Any case with missing data or a missing lable is ignored.
}
