% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tSNE}
\alias{tSNE}
\title{\code{tSNE}}
\usage{
tSNE(data, is.distance = FALSE, perplexity = 10, seed = 1066)
}
\arguments{
\item{data}{A numeric \code{\link{data.frame}} or \code{\link{matrix}} containing the data to be analyzed.}

\item{is.distance}{Whether \code{data} are a distance \code{\link{matrix}} between points or a
numeric \code{\link{data.frame}} of cases by row and features by column.}

\item{perplexity}{The perplexity coefficient which defines the extent of the locality
of the dimension reduction.}

\item{seed}{Random seed.}
}
\description{
Produce a t-Distributed Stochastic Neighbour Embedding.
}
\details{
If input is not a distance matrix, there must be no missing data or duplicates.
}
